Refactor asset using models to be more similar
Required changes to ensure that all models that use assets can be
updated using the same workers/jobs (avoiding having to duplicate
those workers/jobs).

- Move shared `AssetData` logic from `AttachmentData` to shared
`AssetData` concern
- Add `AssetData` to `ImageData`, `ConsultationResponseFormData` and
`CallForEvidenceResponseFormData`
- Add `publicly_visible?`, `deleted?` and `atachable` to `Image`,
`ConsultationResponseForm` and `CallForEvidenceResponseForm`
