+++ app/models/concerns/attachable.rb
@@ -52,6 +52,10 @@ module Attachable
   end
 
   class Null
+    def id
+      nil
+    end
+
     def publicly_visible?
       false
     end
@@ -73,6 +77,10 @@ module Attachable
     end
 
     def unpublished?
+      false
+    end
+
+    def deleted?
       false
     end
   end
