+++ app/models/consultation_response_form_data.rb
@@ -1,14 +1,27 @@
 class ConsultationResponseFormData < ApplicationRecord
+  include AssetData
+
   mount_uploader :file, ResponseDocumentUploader, mount_on: :carrierwave_file
 
   has_one :consultation_response_form
-  has_many :assets,
-           as: :assetable,
-           inverse_of: :assetable
 
   validates :file, presence: true
 
+  def replaced?
+    false
+  end
+
+  def attachable
+    consultation_response_form || Attachable::Null.new
+  end
+
+  def attachments
+    [consultation_response_form || Attachment::Null.new]
+  end
+
   def auth_bypass_ids
+    return [] if attachable.is_a?(Attachable::Null)
+
     [consultation_response_form.consultation_participation.consultation.auth_bypass_id]
   end
